+++ src_bonus/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: lilefebv <lilefebv@student.42lyon.fr>      +#+  +:+       +#+        */
+/*   By: secros <secros@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/26 10:11:21 by secros            #+#    #+#             */
-/*   Updated: 2025/04/26 18:56:09 by lilefebv         ###   ########lyon.fr   */
+/*   Updated: 2025/04/26 21:00:51 by secros           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -41,7 +41,7 @@ int main(int ac, char **av) {
 	int		youaredead_screen = 0;
 	int		skippall = 0;
 
-	int score = 0;
+	size_t score = 0;
 
 	signal(SIGWINCH, window_resize);
 
@@ -143,7 +143,7 @@ int main(int ac, char **av) {
 				}
 				
 				
-				mvwprintw(death_screen, LINES / 2 - num_leaderboard / 2, COLS / 2 - ((8 + ft_get_n_size(score)) / 2), "Score : %d", score);
+				mvwprintw(death_screen, LINES / 2 - num_leaderboard / 2, COLS / 2 - ((8 + ft_get_n_size(score)) / 2), "Score : %zu", score);
 				if (current_player.score < (size_t)score)
 				{
 					mvwprintw(death_screen, LINES / 2 + 1 - num_leaderboard / 2, COLS / 2 - 21 / 2, "New personnal best !");
